Relax dst() equality property
As mentioned in the comment, tracking down all these bugs and
cataloguing them is probably best done elsewhere, and until the three
libraries are smoothed out and at least consistent, it's best to relax
the property.

This still requires some assumptions, because there are cases in all
three libraries where a DST offset returns timedelta(0), but those cases
should be a subset of the inconsistencies where x.dst() == y.dst().