Try to make tz data versions consistent across CI jobs
Apparently pytz only uses the system data if you tell it where it is;
this causes problems because (at the moment) the dateutil time zone data
is not the latest version, so the Python 2.7 jobs are failing.

We'll pin pytz to 2019c on Python 2.7 for now, and try to get pytz to
pick up the system data in all other cases.